requirements.txt
Authlib==0.15.3
cachetools==5.0.0
certifi==2022.12.7
cffi==1.14.5
cfn-flip==1.2.3
chardet==4.0.0
click==8.0.1
cryptography==39.0.1
Deprecated==1.2.12
Flask==2.1.0
flask-feather==0.1.0
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==2.10
importlib-metadata==4.11.3
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.1.2
oauthlib==3.2.2
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyGithub==1.55
PyJWT==2.6.0
PyNaCl==1.4.0
pyparsing==3.0.7
python-dotenv==0.20.0
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.5
Werkzeug==2.2.3
wrapt==1.12.1
zipp==3.7.0
